LP0: vcm30t124: Enable PWR_INT_N
Bug  1553974
Bug  1360277

Patch to enable wake source PWR_INT_N wake source 18 in
PMC. We bypass the interrupt framework, so Wake event is
not serviced on wake from LP0.

The wake button on Jetson is triggered by either pressing
Wake Switch S2 or by FT232H USB-to-UART/FIFO. We see that
the wake button is sometimes kept asserted due to FT232H.
And this used to cause flood of interrupts resulting in
soft lockup. This patch addresses soft lockup issue.

Change-Id: I828beed8f986bab7561042745d4453b8a2769a61
Signed-off-by: Rajesh Hosakote <rhosakote@nvidia.com>
Reviewed-on: http://git-master/r/501275
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Sandeep Trasi <strasi@nvidia.com>