misc: tegra-baseband: enable modem on XHCI HSIC
Enable modem support on both EHCI and XHCI HSIC. To determine which HC
to use, "nvidia,use-xhci-hsic" property in node /modem in DT should be
set to 1 if XHCI is to be used. If the property is absent or it is set
to 0, EHCI will be used.

Bug 200029129

Change-Id: I4b70f6eddd2dc3418a432ea543425b195768de00
Signed-off-by: Mark Kuo <mkuo@nvidia.com>
Reviewed-on: http://git-master/r/499283
Reviewed-by: Neil Patel <neilp@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Steve Lin <stlin@nvidia.com>