tegra: thermal: t210: enable thermal actions in dt
Enable shutdown, heavy throttling and balanced throttling on soctherm
zones via of-thermal device tree configuration per platform.

Presently enabled for ers E2220 and Loki-E platforms.

Note: Since any changes to trip temperatures via sysfs nodes is
handled in of-thermal layer and not propagated to the underlying
thermal-sensor driver, we currently don't support the ability to
change shutdown and hw-throttle trip points. This would be supported
once the required enhancements to of-thermal code is done.

Change-Id: I16d3daf3fcb08eefefbdd565e3025489bdcfeea4
Signed-off-by: Diwakar Tundlam <dtundlam@nvidia.com>
Reviewed-on: http://git-master/r/553056