mmc: host: sdhci: fix sdmmc clock in DDR50 mode
For emmc on vcm30t124, data checksum error is reported during
data transfers in DDR50 mode @ 26Mhz. In this mode the sdmmc
clock (102 MHz) is divided by a factor of 4 to provide 26MHz
to the card.
This is SW WAR that fixes this issue by making sdmmc clock to
51 MHz instead of 102MHz and there by reducing the dividing
factor to 2 instead of 4 to provide same 26MHz clock to card.

Bug 200019683

Change-Id: I95c8e0ea8185eff9c45fa0ffe7fc14d24efe0256
Signed-off-by: kirankumar Bobbu <kbobbu@nvidia.com>
Reviewed-on: http://git-master/r/554550
Reviewed-by: Automatic_Commit_Validation_User
Tested-by: Sandeep Trasi <strasi@nvidia.com>
Reviewed-by: Sandeep Trasi <strasi@nvidia.com>