usb: gadget: xudc: war for disconnect in U2/U3
Issue is that on a disconnect from U3, the SS Link gets stuck in
Resume state. For U2 case, the SS link gets stuck in U2 state.
We will have to do a directed U0 with Remote Wake Enabled to get
the state machine out of these states.

Bug 200041378
Bug 200033743
Bug 200046666

Change-Id: I7395cb570cf3b0df6375d1a29fcacfabcd22f1e1
Signed-off-by: Henry Lin <henryl@nvidia.com>
Reviewed-on: http://git-master/r/554603
Reviewed-by: Ashutosh Jha <ajha@nvidia.com>
Tested-by: Ashutosh Jha <ajha@nvidia.com>