mmc: increase timeout for MMC COMBO commands
After all COMBO commands are executed we send CMD13
to check for card status if it is moved from PROGRAMMING state.
The max wait timeout was too less for refresh commands.

We need have to wait for minium of 5 sec, becasue of usleep_range
this range can be from 5 to 25 Secs

Bug 200045977
Bug 200044665
Bug 200026104
Bug 1501708

Change-Id: Ic6e91ed48a6e67968ddda0186c4659a1ebc57015
Signed-off-by: Seshagiri.H <sholi@nvidia.com>
Reviewed-on: http://git-master/r/556608
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sandeep Trasi <strasi@nvidia.com>