arm: tegra: mc: Parse DT prop
Parse a dt property that until now was being ignored. This property
specifies the number of interrupt bits in the MC and is used for
limiting a for loop that prints out errors seen on all the MC
clients. When the interrupt count was 0, it appeared as though the
MC was not recording any errors.

Bug 200046934

Change-Id: I266e3b1eef62219be9cb80541944b83073dbeb2a
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/560120
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>