regmap: cache_rbtree: ignore memleak reports for allocated memory for cache
Some of driver like PMIC driver uses the RBTREE cache mechanism for caching the
registers. All non-volatile registers are cache enabled and some of the cached
registered only accessed once on whole system.

kmemleak debug utility reports the allocated memory to be leak if it is not get
accessed for 10 min and on this case, above registers which is accessed during
boot for which cache memory is allocated reported as mem-leak as it is not accessed
after boot.

Ignoring such allocation as it is known that it is not leak to avoid false report.

bug 1569775

Change-Id: Ia8eb2c76f0d841496be0cfbda579888cc72e2240
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/590358