video: tegra: host: fix syncpt wait for virt case
The "condition" argument to the wait_event_* APIs is
syncpt_update_min_is_expired(), which reads the syncpoint value
from HW then checks if the threshold is expired. In the
virtualization case, the syncpoint read requires an inter-VM
switch to the server. This results in the wait_event_* APIs
never actually blocking, as evaluating "condition" causes
current->state to change from  TASK_INTERRUPTIBLE to
TASK_RUNNING.

To resolve this, the current min syncpoint value is used for the
comparison. This is OK since it will have been updated by the
syncpoint intr thread responsible for waking up the wait queue.

Bug 200048722

Change-Id: Ic55a96b467289fa26a608205f325778fb34291f9
Signed-off-by: Aingara Paramakuru <aparamakuru@nvidia.com>
Reviewed-on: http://git-master/r/592825
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Arto Merilainen <amerilainen@nvidia.com>