arm64: tegra21: emc: modify EMC driver init
Change the EMC driver init so that the emc timers - thermal and
periodic training for t21x - happen as a subcall of the general
EMC probe. One caveat is that the thermal device registration
has to be delayed because the EMC driver init runs before the
thermal framework seems to exist.

Change-Id: I7184dcd8629cbf29547262418a6a283de6ef5b26
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/660451
(cherry picked from commit 5922fd8b60ae4d197925dd6142fdc1a7799102e1)
Reviewed-on: http://git-master/r/663395
Reviewed-by: Mitch Luban <mluban@nvidia.com>