HACK: iommu/arm-smmu: fix swgroup configuration
The param "ASID" is numerically equal to the stream-ID. The cgid (for
nvhost this value is 22) has a one-to-one mapping to swid (for nvhost
cgidtoswid[22] = 6). Based on this value  the address of in
MS_SMMU_<swgroup>_ASID_0 is determined. (which is 0x22c1_0250). This
contains the value of the stream-ID to be passed to the SMMU Fastmodel.

We use swgroupID == streamID currently.

Change-Id: I5447b9b4fc76e872fcd43a7a52eba0c508503d83
Signed-off-by: Hiroshi Doyu <hdoyu@nvidia.com>
Reviewed-on: http://git-master/r/595780
Reviewed-by: Sri Krishna Chowdary <schowdary@nvidia.com>
GVS: Gerrit_Virtual_Submit