clock: tegra: fix typecast
Remove unnecessary typecast which actually causes Sparse warning.

The "uintptr_t" typecast is only needed when there is a need to
temporarily save pointer to an integer which is later converted
to pointer again. Here the pointer is directly used for MMIO access,
therefore there is no need to do extra typecasting.

bug 200032218

Change-Id: I4476aa8cabfba05315d7b027e385bdd4a8c547de
Signed-off-by: Bo Yan <byan@nvidia.com>
Reviewed-on: http://git-master/r/598937
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Tested-by: Sachin Nikam <snikam@nvidia.com>