platform: tegra: reboot: register system restart handler
In order to restart the system in a platform specified way, instead
of depending on PSCI, register a restart handler with a higher priority.
The priority is kept higher than the default to let this handler execute
system restart. If we are unable to do so, then the default handler is
given a chance to restart the system.

Change-Id: I445d4c77c85950baf73688bf31ade9a36ba9ce48
Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>
Reviewed-on: http://git-master/r/606084
GVS: Gerrit_Virtual_Submit