Revert "video: tegra: host: Rework serialization"
This reverts commit fb744620d622ba4d39ca7125d81ad80f88cdba2f.

The commit added a "job tracking" syncpoint to ensure that we
release job resources after handling the whole work (including
added opcodes in the end) before turning off the engine.

The new increment was accounted as a job increment and therefore
the number of job increments increased by one. However, this
caused a regression to use cases where the userspace had strict
assumption that the syncpoint range was from fence-num_incrs up
to fence.

This patch modifies the code so that the added job tracking increment
is not accounted to the fences that we return back to userspace.

Bug 1576399

Change-Id: I1bc8385636654c6e7be5c668d426cc9bcced9951
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/606879