tegra: clocks: Parse dfll_boot_khz kernel cmd
1.
Parse dfll_boot_khz kernel command line parameter
in early init to initialize dfll frequency clock
rate varaible.

This is required to make sure that SW clock rate
variable to be in sync with HW dfll freq rate and
for not to have sw cpu feq as zero.

With this change, below issue won't occur:

In a Virtual guest platform, where kernel is
not required to do any clock operations on CPU,
existing frequency in cpufreq coming a zero.

2.
Add appropriate check to cl_dvfs APIs for
not to invoke when not initialized

Bug 200052836

Change-Id: Idf102c5f27f9c2c623075c55f01adf2baf7e7c96
Signed-off-by: sreenivasulu velpula <svelpula@nvidia.com>
Reviewed-on: http://git-master/r/656289
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
Reviewed-by: Sandeep Trasi <strasi@nvidia.com>