usb: gadget: xudc: add delay before ringing DB
When we have a large number of endpoints configured (more than 10),
on U3 exit, SW would clear all the Pause bits that was set as a result of
the U3 entry and will ring doorbell to start transfer on some endpoints.
If the time difference between SW clearing the Pause bit and SW ringing
doorbell was very little (less than 200ns), this could lead to the doorbell
getting dropped and the corresponding transfer not completing.
Workaround is to add 500ns wait.

Bug 1584683

Change-Id: I1355c2ea20a2864ed9be98b8d02a08274742a2a1
Signed-off-by: Hui Fu <hfu@nvidia.com>
Reviewed-on: http://git-master/r/658564
Reviewed-by: Ashutosh Jha <ajha@nvidia.com>
Tested-by: Ashutosh Jha <ajha@nvidia.com>
Reviewed-on: http://git-master/r/664643