tegra: watchdog: add virtual watchdog driver
Add a minimal virtual watchdog driver (TEGRA_HV_WATCHDOG) for use
on hypervisor-enabled Tegra systems, and enable it by default for
vcm30t124 builds. When enabled, the driver communicates refresh
events at fixed intervals to a monitor service running in a
different VM. Currently, the refresh interval is not configurable;
that ability will be added as a future change.

Bug 1568715

Change-Id: I6415f33fc885d12b53544292e7a44b87612fe401
Signed-off-by: Dennis Kou <dkou@nvidia.com>
Reviewed-on: http://git-master/r/660403
Reviewed-by: Alexander Van Brunt <avanbrunt@nvidia.com>
Tested-by: Alexander Van Brunt <avanbrunt@nvidia.com>