ASoC: ad193x: Keep DAC output stage active in idle
Setting the DAC power-down bit for the ad193x will also disable
the DAC output amplifier. This will cause audible clicks and pops
when starting or stopping playback. To prevent this a new widget is
introduced that controls the DAC power-down bit. This widget is
connected to both the DAC and a newly introduced VMID widget. This
makes sure that the DAC power-down bit is not set as long as an
audio sink is connected to the DAC output. At the same time the
PLL and SYSCLK will still be disabled when no playback or capture
stream is active.

Bug 1506525

(cherrypicked from upstream commit 1b86a3fa4eb3c7a6d738fa21475b92493f8952b1)

Change-Id: I473483c84f5d07b3925b6d4b5585549aa3d45617
Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Signed-off-by: Mark Brown <broonie@kernel.org>
Signed-off-by: Arun Shamanna Lakshmi <aruns@nvidia.com>
Reviewed-on: http://git-master/r/665271
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Brian Liu <briliu@nvidia.com>
Reviewed-by: Uday Gupta <udayg@nvidia.com>
Tested-by: Uday Gupta <udayg@nvidia.com>
Reviewed-by: Nitin Pai <npai@nvidia.com>
GVS: Gerrit_Virtual_Submit